Moving render graph conversions to same file as other renderer type conversions
