--- magnolia/src/private/render_graph_type_conversions.hpp
@@ -1,9 +1,0 @@
-#pragma once
-
-#include <vulkan/vulkan_handles.hpp>
-
-namespace mag
-{
-    enum class AttachmentState;
-    vk::AttachmentLoadOp mag_attachment_state_to_vk(const AttachmentState state);
-};  // namespace mag
